% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hefi2019.score.macro.R
\name{hefi2019}
\alias{hefi2019}
\title{Healthy Eating Food Index-2019 Scoring Algorithm}
\usage{
hefi2019(
  indata,
  vegwfruit,
  wholegrfoods,
  nonwholegrfoods,
  profoodsanimal,
  profoodsplant,
  otherfoods,
  water_and_other_healthy,
  unsweetmilk,
  unsweetplantbevpro,
  otherbev,
  mufat,
  pufat,
  satfat,
  sugars,
  kcal,
  sodium
)
}
\arguments{
\item{indata}{Input dataset with dietary constituents}

\item{vegwfruit}{RAs from vegetables and (whole) fruits}

\item{wholegrfoods}{RAs from whole-grain foods}

\item{nonwholegrfoods}{RAs from non-whole grain foods}

\item{profoodsanimal}{RAs from animal-based protein foods}

\item{profoodsplant}{RAs from plant-based protein foods}

\item{otherfoods}{RAs from all others foods (i.e., not considered in the variables above)}

\item{water_and_other_healthy}{Grams of water and other healthy beverages (see definition in Lamarche et al. 2021)}

\item{unsweetmilk}{Grams of unsweetened milk (all \% M.F.)}

\item{unsweetplantbevpro}{Grams of unsweetened plant-based beverages}

\item{otherbev}{Grams of all other beverages (artificially- or sugar-sweetened beverages, juices, sweetened milk or plant-based beverages. See definition in Lamarche et al. 2021)}

\item{mufat}{Grams of fat from monounsaturated fats}

\item{pufat}{Grams of fat from polyunsaturated fats}

\item{satfat}{Grams of fat from saturated fats}

\item{sugars}{Grams of free sugars}

\item{kcal}{Total energy intake, kcal}

\item{sodium}{Milligrams of sodium}
}
\value{
Input dataset (\code{indata}) with additional variables including density of intakes (i.e., ratios of dietary constituents), total score and component subscores
}
\description{
This function scores dietary constituents provided in the input dataset
according to the Healthy Eating Food Index (HEFI)-2019 scoring algorithm
(Lamarche et al., 2021; Brassard et al., 2021). The original variables are
kept in the output data. New variables include density of intakes
(i.e., ratios of dietary constituents), the total HEFI-2019 and its
components subscores.
}
