% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hefi2019.scoring.R
\name{hefi2019}
\alias{hefi2019}
\title{Healthy Eating Food Index-2019 Scoring Algorithm}
\usage{
hefi2019(
  indata,
  vegfruits,
  wholegrfoods,
  nonwholegrfoods,
  profoodsanimal,
  profoodsplant,
  otherfoods,
  waterhealthybev,
  unsweetmilk,
  unsweetplantbevpro,
  otherbeverages,
  mufat,
  pufat,
  satfat,
  freesugars,
  sodium,
  energy
)
}
\arguments{
\item{indata}{Input data frame with dietary constituents}

\item{vegfruits}{Reference amounts (RA) from vegetables and fruits (excludes fruit juices)}

\item{wholegrfoods}{RA from whole-grain foods (includes whole wheat)}

\item{nonwholegrfoods}{RA from non-whole grain foods}

\item{profoodsanimal}{RA from animal-based protein foods (excludes milk)}

\item{profoodsplant}{RA from plant-based protein foods (excludes plant-based beverages)}

\item{otherfoods}{RA from all other foods not considered in the categories above (vegetables and fruits, grain foods, protein foods), excludes beverages}

\item{waterhealthybev}{Grams of water and other healthy beverages not in unsweetened milk or unsweetened plant-based beverages with sufficient protein}

\item{unsweetmilk}{Grams of unsweetened milk (includes all \% M.F.)}

\item{unsweetplantbevpro}{Grams of unsweetened plant-based beverages with sufficient protein}

\item{otherbeverages}{Grams of all other beverages (artificially- or sugar-sweetened beverages, juices, sweetened milk or plant-based beverages, etc.)}

\item{mufat}{Grams of fat from monounsaturated fats}

\item{pufat}{Grams of fat from polyunsaturated fats}

\item{satfat}{Grams of fat from saturated fats}

\item{freesugars}{Grams of free sugars}

\item{sodium}{Milligrams of sodium}

\item{energy}{Total energy intake, kcal}
}
\value{
A data frame (\code{indata}) with additional variables including:
\itemize{
\item Density of intakes (i.e., ratios of dietary constituents)
\item HEFI-2019 total score
\item HEFI-2019 component scores
}
}
\description{
This function scores dietary constituents provided in the input data frame
according to the Healthy Eating Food Index (HEFI)-2019 scoring algorithm.
The original variables are kept in the output data frame.
}
\details{
When no foods, beverages or energy are reported, ratios are
not calculated and a score of \code{0} is assigned to the corresponding
components. Missing data for any dietary constituents will results in
missing components score(s) and total score.

Suggested layout for the input data frame:
the function should ideally be applied to a data frame where rows correspond
to individuals and dietary constituents are columns.

Caution variable names:
\code{totfoodsRA}, \code{RATIO_VF}, \code{RATIO_WGTOT}, \code{RATIO_WGGR},
\code{RATIO_PRO}, \code{RATIO_PLANT}, \code{RATIO_FA}, \code{RATIO_BEV},
\code{SFA_PERC}, \code{SUG_PERC}, \code{SODDEN} are reserved for this function.
}
\examples{
\dontrun{
mydata_scored <-
hefi2019(indata             = mydata,
         vegfruits          = "RA_vegfruits",
         wholegrfoods       = "RA_wholegrfoods",
         nonwholegrfoods    = "RA_nonwgfoods",
         profoodsanimal     = "RA_profoodsanimal",
         profoodsplant      = "RA_profoodsplant",
         otherfoods         = "RA_otherfoods",
         waterhealthybev    = "G_waterhealthybev",
         unsweetmilk        = "G_milk",
         unsweetplantbevpro = "G_plantbevpro",
         otherbeverages     = "G_otherbeverages" ,
         mufat              = "G_mufa" ,
         pufat              = "G_pufa" ,
         satfat             = "G_sfa" ,
         freesugars         = "G_freesugars",
         sodium             = "MG_sodium",
         energy             = "energy")
}

}
\references{
Brassard et al. Appl Physiol Nutr Metab. 2022. Development of the Healthy Eating Food Index (HEFI)-2019 measuring adherence to Canada's Food Guide 2019 recommendations on healthy food choices. \url{https://doi.org/10.1139/apnm-2021-0415}

Brassard et al. Appl Physiol Nutr Metab. 2022. Evaluation of the Healthy Eating Food Index (HEFI)-2019 measuring adherence to Canada's Food Guide 2019 recommendations on healthy food choices. \url{https://doi.org/10.1139/apnm-2021-0416}
}
\seealso{
Health Canada. Open Government Portal. 2023. The Healthy Eating Food Index 2019. Record ID: 29892c85-2ff5-484c-873c-f494ffba6e1b. \url{https://open.canada.ca/data/en/dataset/29892c85-2ff5-484c-873c-f494ffba6e1b}
}
