% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hefi2019.scoring.macro.R
\name{hefi2019}
\alias{hefi2019}
\title{Healthy Eating Food Index-2019 Scoring Algorithm}
\usage{
hefi2019(
  indata,
  vegfruits,
  wholegrfoods,
  nonwholegrfoods,
  profoodsanimal,
  profoodsplant,
  otherfoods,
  waterhealthybev,
  unsweetmilk,
  unsweetplantbevpro,
  otherbeverages,
  mufat,
  pufat,
  satfat,
  freesugars,
  sodium,
  energy
)
}
\arguments{
\item{indata}{Input data set with dietary constituents}

\item{vegfruits}{Reference amounts (RA) from vegetables and fruits (excludes fruit juices)}

\item{wholegrfoods}{RA from whole-grain foods (includes whole wheat)}

\item{nonwholegrfoods}{RA from non-whole grain foods}

\item{profoodsanimal}{RA from animal-based protein foods (excludes milk)}

\item{profoodsplant}{RA from plant-based protein foods (excludes plant-based beverages)}

\item{otherfoods}{RA from all other foods not considered in the categories above (vegetables and fruits, grain foods, protein foods), excludes beverages}

\item{waterhealthybev}{Grams of water and other healthy beverages not in unsweetened milk or unsweetened plant-based beverages with sufficient protein}

\item{unsweetmilk}{Grams of unsweetened milk (includes all \% M.F.)}

\item{unsweetplantbevpro}{Grams of unsweetened plant-based beverages with sufficient protein}

\item{otherbeverages}{Grams of all other beverages (artificially- or sugar-sweetened beverages, juices, sweetened milk or plant-based beverages, etc.)}

\item{mufat}{Grams of fat from monounsaturated fats}

\item{pufat}{Grams of fat from polyunsaturated fats}

\item{satfat}{Grams of fat from saturated fats}

\item{freesugars}{Grams of free sugars}

\item{sodium}{Milligrams of sodium}

\item{energy}{Total energy intake, kcal}
}
\value{
Input data set (\code{indata}) with additional variables including density of intakes (i.e., ratios of dietary constituents), total and component scores
}
\description{
This function scores dietary constituents provided in the input data set
according to the Healthy Eating Food Index (HEFI)-2019 scoring algorithm
See Brassard et al., Appl Physiol Nutr Metab 2022 for details and
information on classification of foods.
The original variables are kept in the output data. The new variables include
density of intakes (i.e., ratios of dietary constituents), the total
HEFI-2019 score and component scores.
Of note, when no foods, beverages or energy are reported, ratios are
not calculated and a score of 0 is assigned to the corresponding
components. Missing data for any dietary constituents will results in
missing components score(s) and total score.
}
\details{
Suggested layout for the input data set:
the function should ideally be applied to a data set where rows correspond
to individuals and dietary constituents are columns.

Caution:  variable names "unsweetmilk_RA", "unsweetplantbevpro_RA",
"totfoodsRA", "totgrain", "totpro", "totbev", "unsatfat",
"RATIO_VF", "RATIO_WGTOT", "RATIO_WGGR", "RATIO_PRO",
"RATIO_PLANT", "RATIO_FA", "RATIO_BEV", "SFA_PERC", "SUG_PERC",
"SODDEN", "FATmin", "FATmax", "SFAmin", "SFAmax", "SUGmin",
"SUGmax", "SODmin", "SODmax" are reserved for this function.
}
\references{
Brassard et al. Appl Physiol Nutr Metab. 2022. Development of the Healthy Eating Food Index (HEFI)-2019 measuring adherence to Canada's Food Guide 2019 recommendations on healthy food choices
}
