% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hefi2019.score.macro.R
\name{hefi2019}
\alias{hefi2019}
\title{Healthy Eating Food Index-2019 Scoring Algorithm}
\usage{
hefi2019(
  indata,
  vegwfruits,
  wholegrfoods,
  nonwholegrfoods,
  profoodsanimal,
  profoodsplant,
  otherfoods,
  water_and_other_healthy,
  unsweetmilk,
  unsweetplantbevpro,
  otherbev,
  mufat,
  pufat,
  satfat,
  sugars,
  kcal,
  sodium
)
}
\arguments{
\item{indata}{Input dataset with dietary constituents}

\item{vegwfruits}{RAs* from vegetables and fruits (excludes fruit juices)}

\item{wholegrfoods}{RAs from whole-grain foods}

\item{nonwholegrfoods}{RAs from non-whole grain foods}

\item{profoodsanimal}{RAs from animal-based protein foods}

\item{profoodsplant}{RAs from plant-based protein foods}

\item{otherfoods}{RAs from all others foods (i.e., not considered in the variables above)}

\item{water_and_other_healthy}{Grams of water and other healthy beverages}

\item{unsweetmilk}{Grams of unsweetened milk (all \% M.F.)}

\item{unsweetplantbevpro}{Grams of unsweetened plant-based beverages}

\item{otherbev}{Grams of all other beverages (artificially- or sugar-sweetened beverages, juices, sweetened milk or plant-based beverages.)}

\item{mufat}{Grams of fat from monounsaturated fats}

\item{pufat}{Grams of fat from polyunsaturated fats}

\item{satfat}{Grams of fat from saturated fats}

\item{sugars}{Grams of free sugars}

\item{kcal}{Total energy intake, kcal}

\item{sodium}{Milligrams of sodium}
}
\value{
Input dataset (\code{indata}) with additional variables including density of intakes (i.e., ratios of dietary constituents), total and component scores
}
\description{
This function scores dietary constituents provided in the input dataset
according to the Healthy Eating Food Index (HEFI)-2019 scoring algorithm
See Brassard et al., Appl Physiol Nutr Metab 2022 for details and
information on classification of foods.
The original variables are kept in the output data. New variables include
density of intakes (i.e., ratios of dietary constituents), the total
HEFI-2019 score and its components scores.
Of note, when no foods, beverages or energy is reported, ratios are
not calculated and a score of 0 is assigned to the corresponding
components.
}
\details{
(*) RAs indicate reference amounts (Brassard et al., Appl Physiol Nutr Metab 2022).
Caution:  variable names "unsweetmilk_RA", "unsweetplantbevpro_RA",
"totfoodsRA", "totgrain", "totpro", "totbev", "unsatfat",
"RATIO_VF", "RATIO_WGTOT", "RATIO_WGGR", "RATIO_PRO",
"RATIO_PLANT", "RATIO_FA", "RATIO_BEV", "SFA_PERC", "SUG_PERC",
"SODDEN", "FATmin", "FATmax", "SFAmin", "SFAmax", "SUGmin",
"SUGmax", "SODmin", "SODmax" are reserved for this function.
}
